# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Export Service API",
    "version": "1.0.0",
    "description": "API documentation for Export Job System"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/download/{id}": {
      "get": {
        "summary": "Download exported file",
        "description": "Downloads the generated export file for the given export job. The download link remains valid only for a limited time (e.g., 1 hour) after the export job is completed.",
        "tags": ["Exports"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Export job identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export file downloaded successfully",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Export job not found"
          },
          "410": {
            "description": "Download link has expired"
          }
        }
      }
    }
  },
...
}